pfind: Improve `find -> ls` logic
Use `xargs` to pass output of `find` to arguments of `ls`, instead of
`mapfile` and a bash array.
Not only is this a more efficient method, it also has improved behavior.
`mapfile` would choke on filenames containing newlines, but `xargs` does
not.  As a bonus, the code is somewhat less cryptic-looking.